Update UI theme and enhance Search functionality
- Implement `Inter` font family and update Typography with Material3 styles
- Refactor `SearchItem` to use a generic `SearchBaseCard` for movies, series, and persons
- Rename `SerialSearchResult` to `SeriesSearchResult` across domain and UI layers
- Add `formatDate` utility and `SearchItemPreviewProvider` for better UI previews
- Update `SearchScreen` with `AnimatedContent` transitions and new `Empty` and `Waiting` states
- Adjust `SearchViewModel` to handle short queries and empty paging data
- Fix `KnownFor` ID type from `Long` to `Int` in domain and DTO models

+++ feature/tmdb/src/main/java/com/vroff/tmdb/entity/search/KnownForDTO.kt
@@ -1,13 +1,18 @@
 package com.vroff.tmdb.entity.search
 
 import com.google.gson.annotations.SerializedName
+import com.vroff.domain.model.BackdropImage
+import com.vroff.domain.model.PosterImage
+import com.vroff.domain.model.ProfileImage
+import com.vroff.domain.model.tmdb.search.Gender
 import com.vroff.domain.model.tmdb.search.KnownFor
+import com.vroff.domain.model.tmdb.search.MediaType
 
 data class KnownForDTO(
     val adult: Boolean,
     @SerializedName("backdrop_path")
     val backdropPath: String?,
-    val id: Long,
+    val id: Int,
     val name: String?,
     @SerializedName("original_name")
     val originalName: String?,
